Move UserDefaultStringTests to its own test file
